integration of boardEditor and gameEditor menu
refactoring, and integration of the gameEditor frontend components
