+++ src/GameEditor/Controllers/GameEditorMainController.java
@@ -14,6 +14,7 @@ import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.scene.control.Button;
 import javafx.scene.control.ComboBox;
+import javafx.scene.control.Tab;
 import javafx.scene.input.MouseEvent;
 import javafx.scene.layout.HBox;
 import javafx.scene.text.TextFlow;
@@ -170,5 +171,14 @@ public class GameEditorMainController {
         else if (tokenType == ""){}
         else {}
     }
+
+    @FXML
+    //controller method to add BoardEditor node to corresponding tab in GameEditorMain
+    private Tab boardEditor;
+
+
+
+
+
 
 }
