+++ src/GameEditor/Views/GameEditorMain.fxml
@@ -10,9 +10,10 @@
 <top>
     <TabPane minHeight="410.0" minWidth="700.0" prefHeight="200.0" prefWidth="200.0" tabClosingPolicy="UNAVAILABLE" tabMaxHeight="50.0" tabMaxWidth="118.0" tabMinHeight="50.0" tabMinWidth="118.0" BorderPane.alignment="CENTER">
         <tabs>
-            <Tab text="Board Editor">
+            <Tab text="Board Editor" fx:id="boardEditor">
                 <content>
-                    <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="180.0" prefWidth="200.0" />
+                    <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="180.0" prefWidth="200.0">
+                    </AnchorPane>
                 </content>
             </Tab>
             <Tab text="Rule Editor">
